fix: 优化 模板 原样输出功能

+++ src/document/Parser/utils/Community.php
@@ -29,7 +29,7 @@ class Community
         }
 
         // Fallback (PHP 4.2+)
-        mt_srand((double) microtime() * 10000);
+        mt_srand((int) microtime() * 10000);
         $charid = strtolower(md5(uniqid(rand(), true)));
         $hyphen = chr(45);                  // "-"
         $lbrace = $trim ? "" : chr(123);    // "{"
